Cleanup + Formatting
