+++ .gitignore
@@ -14,3 +14,4 @@ MapleWebServer/Data
 workspace.code-workspace
 .env
 MapleServer2/PacketStructures
+MapleServer2.sln.DotSettings.user
