+++ MapleServer2/Constants/SendOp.cs
@@ -21,8 +21,8 @@ public enum SendOp : ushort
     SYNC_NUMBER = 0x0014,
     SERVER_ENTER = 0x0015,
     REQUEST_FIELD_ENTER = 0x0016,
-    FIELD_ADD_USER = 0x0017,
-    FIELD_REMOVE_USER = 0x0018,
+    FIELD_ADD_PLAYER = 0x0017,
+    FIELD_REMOVE_PLAYER = 0x0018,
     FIELD_ENTRANCE = 0x0019,
     ROOM_DUNGEON = 0x001A,
     RESULTS = 0x001B,
@@ -123,7 +123,7 @@ public enum SendOp : ushort
     FITTING_DOLL = 0x007D,
     BONUS_GAME = 0x007E,
     LOAD_UGC_MAP = 0x007F,
-    PROXY_GAME_OBJ = 0x0080,
+    FIELD_OBJECT = 0x0080,
     GEM = 0x0081,
     TAXI = 0x0082,
     FIND_FIELDS = 0x0083,
