+++ MapleServer2/Packets/RequestFieldEnterPacket.cs
@@ -1,0 +1,24 @@
+﻿using MaplePacketLib2.Tools;
+using MapleServer2.Constants;
+using MapleServer2.Types;
+
+namespace MapleServer2.Packets;
+
+public static class RequestFieldEnterPacket
+{
+    public static PacketWriter RequestEnter(IFieldActor<Player> fieldPlayer)
+    {
+        PacketWriter pWriter = PacketWriter.Of(SendOp.REQUEST_FIELD_ENTER);
+        pWriter.WriteByte(0x00);
+        pWriter.WriteInt(fieldPlayer.Value.MapId);
+        pWriter.WriteByte();
+        pWriter.WriteByte();
+        pWriter.WriteInt();
+        pWriter.WriteInt();
+        pWriter.Write(fieldPlayer.Coord);
+        pWriter.Write(fieldPlayer.Rotation);
+        pWriter.WriteInt(); // Whatever is here seems to be repeated by client in FIELD_ENTER response.
+
+        return pWriter;
+    }
+}
